完成安装程序

+++ app/Http/Controllers/InstallController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\ResponseHttpException;
 use Illuminate\Http\Request;
 
 /**
@@ -14,21 +15,53 @@ class InstallController extends Controller
     //安装首页
     public function index()
     {
-        return $this->migrate();
-//        try {
+        return view('install.index');
+    }
 
-//        } catch (\Exception $exception) {
-//            dd($exception->getMessage());
-//        }
+    public function database()
+    {
+        return view('install.database');
+    }
+
+    public function connect(Request $request)
+    {
+        try {
+            $host = $request->input('host');
+            $db = $request->input('database');
+            $user = $request->input('username');
+            $pass = $request->input('password');
+            new \PDO("mysql:host={$host};dbname={$db}", $user, $pass);
+            put_contents_file(base_path('database.php'),
+                array_merge(array_merge(include base_path('database.php'), $request->except('_token'))));
+            return redirect()->route('install.migrate')->with('success', '数据库配置正确');
+        } catch (\Exception $e) {
+            return redirect(route('install.database'))->with('info', '数据为连接失败');
+        }
     }
 
     public function migrate()
     {
-        \Artisan::call('module:migrate-reset');
-        \Artisan::call('migrate:reset');
-        \Artisan::call('migrate', ['--seed' => null]);
-        \Artisan::call('module:migrate');
-//        \Artisan::call('passport:client', ['--password' => true]);
-        return response()->json(['code' => 0, 'message' => "安装表创建成功"]);
+        return view('install.migrate');
+    }
+
+    public function create()
+    {
+        try {
+            if (\Schema::hasTable('migrations')) {
+                \Artisan::call('module:migrate-reset');
+                \Artisan::call('migrate:reset');
+            }
+            \Artisan::call('migrate', ['--seed' => null]);
+            \Artisan::call('module:migrate');
+            return redirect()->route('install.complete')->with('success', '数据表创建成功');
+        } catch (\Exception $e) {
+            throw new ResponseHttpException('数据表创建失败' . $e->getMessage());
+        }
+    }
+
+    public function complete()
+    {
+        file_put_contents(base_path('install.lock'), 'The installation is complete');
+        return view('install.complete');
     }
 }
